requirements.txt
asgiref==3.2.3
asn1crypto==0.24.0
boto==2.49.0
boto3==1.7.76
botocore==1.10.76
braintree==3.46.0
certifi==2018.4.16
cffi==1.14
chardet==3.0.4
cryptography==2.2.2
dj-database-url==0.5.0
Django==2.1.5
django-appconf==1.0.3
django-crispy-forms==1.7.2
django-filter==2.2.0
django-pagedown==1.0.6
django-registration-redux==1.10
django-select2==6.0.0
django-storages==1.6.6
djangorestframework==3.11.0
docutils==0.14
enum34==1.1.6
geoip2==3.0.0
gunicorn==19.9.0
html5lib==1.0.1
idna==2.7
ipaddress==1.0.22
jmespath==0.9.3
Markdown==3.0.1
maxminddb==1.5.2
Pillow==6.2.1
psycopg2-binary
pycparser==2.18
pyOpenSSL==18.0.0
PyPDF2==1.26.0
python-dateutil==2.7.3
pytz==2018.5
reportlab==3.5.34
requests==2.22.0
s3transfer==0.1.13
six==1.11.0
sqlparse==0.3.0
stripe==2.42.0
urllib3==1.25.8
webencodings==0.5.1
xhtml2pdf==0.2.4
